+++ packages/core/navigation-next/src/components/Group/index.js
@@ -1,16 +1,24 @@
 // @flow
 
-import React, { Fragment } from 'react';
+import React, { Component, Fragment } from 'react';
 
 import { GroupHeading, Separator } from '../../';
 import type { GroupProps } from './types';
 
-export default ({ children, hasSeparator, heading }: GroupProps) => {
-  return React.Children.count(children) ? (
-    <Fragment>
-      {heading && <GroupHeading>{heading}</GroupHeading>}
-      {children}
-      {hasSeparator && <Separator />}
-    </Fragment>
-  ) : null;
-};
+export default class Group extends Component<GroupProps> {
+  static defaultProps = {
+    hasSeparator: false,
+  };
+
+  render() {
+    const { children, hasSeparator, heading } = this.props;
+
+    return React.Children.count(children) ? (
+      <Fragment>
+        {heading && <GroupHeading>{heading}</GroupHeading>}
+        {children}
+        {hasSeparator && <Separator />}
+      </Fragment>
+    ) : null;
+  }
+}