+++ packages/core/navigation-next/src/components/LayoutManager/types.js
@@ -2,12 +2,19 @@
 
 import type { ComponentType, Node } from 'react';
 
-import { UIState } from '../../';
+import UIController from '../../ui-controller/UIController';
 
-export type LayoutManagerProps = {
+export type ConnectedLayoutManagerProps = {
+  /** Your page content. */
   children: Node,
+  /** A component which will render the container navigation layer. */
   containerNavigation: ?ComponentType<{}>,
+  /** A component which will render the global navigation bar. */
   globalNavigation: ComponentType<{}>,
-  navigationUI: UIState,
+  /** A component which will render the product navigation layer. */
   productNavigation: ComponentType<{}>,
+};
+
+export type LayoutManagerProps = ConnectedLayoutManagerProps & {
+  navigationUIController: UIController,
 };