+++ packages/editor/editor-core/src/plugins/insert-block/ui/ToolbarInsertBlock/index.tsx
@@ -16,6 +16,7 @@ import LinkIcon from '@atlaskit/icon/glyph/editor/link';
 import EmojiIcon from '@atlaskit/icon/glyph/editor/emoji';
 import DateIcon from '@atlaskit/icon/glyph/editor/date';
 import PlaceholderTextIcon from '@atlaskit/icon/glyph/media-services/text';
+import LayoutTwoEqualIcon from '@atlaskit/icon/glyph/editor/layout-two-equal';
 import HorizontalRuleIcon from '@atlaskit/icon/glyph/editor/horizontal-rule';
 import {
   EmojiId,
@@ -428,7 +429,7 @@ export default class ToolbarInsertBlock extends React.PureComponent<
         value: { name: 'layout' },
         tooltipDescription: 'Insert columns',
         tooltipPosition: 'right',
-        elemBefore: <PlaceholderTextIcon label="Insert columns" />,
+        elemBefore: <LayoutTwoEqualIcon label="Insert columns" />,
       });
     }