+++ packages/editor/renderer/src/ui/Renderer/style.tsx
@@ -285,4 +285,23 @@ export const Wrapper: ComponentClass<Props & HTMLAttributes<{}>> = styled.div`
       margin-top: 0;
     }
   }
+
+  & [data-layout-type] {
+    display: none;
+  }
+
+  & [data-layout-type="two_equal"] {
+    display: flex;
+    flex-direction: row;
+    position: relative;
+    width: calc(100% + 48px);
+    left: -28px;
+
+    & > * {
+      margin: ${akGridSizeUnitless}px;
+      padding: ${akGridSizeUnitless * 2}px;
+      flex: 1;
+      min-width: 0;
+    }
+  }
 `;